% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kde.R
\name{kde}
\alias{kde}
\title{Kernel Density Estimation}
\usage{
kde(points, band_width, cell_size, kernel = "quartic", scaled = FALSE, grid)
}
\arguments{
\item{points}{\code{\link[sf]{sf}} \code{data.frame} containing only POINTS.}

\item{band_width}{\code{numeric} specifing the band width for KDE.}

\item{cell_size}{\code{numeric} specifing the distance for equal spaced points or cells. Must be
higher than 0. Can be left out if \code{grid} is provided.}

\item{kernel}{\code{character} specifing type of kernel to use. Available implemented kernels are
\code{"uniform", "quartic", "triweight", "epanechnikov"}. Default is \code{"quartic"} and if
uknown kernel name is used it falls back to the default value.}

\item{scaled}{\code{logical} specifing if the output values should be scaled. Default value is
\code{FALSE}.}

\item{grid}{\code{\link[sf]{sf}} \code{data.frame} outcome of function
\code{\link{create_grid_rectangular}} or \code{\link{create_grid_hexagonal}}.}
}
\value{
\code{\link[sf]{sf}} \code{data.frame}.
}
\description{
KDE for spatial data. The algorithm is heavily inspired by
\href{https://github.com/qgis/QGIS/blob/b3d2619976a69d7fb67b884492da491dfaba287c/src/analysis/raster/qgskde.cpp}{Heatmap tool}
in QGIS. The tutorial is provided \href{https://grindgis.com/software/heat-map-using-qgis}{here}.
}
