% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_points_grid.R
\name{create_raster_rectangular}
\alias{create_raster_rectangular}
\alias{create_raster_hexagonal}
\title{Create raster}
\usage{
create_raster_rectangular(
  geometry,
  cell_size,
  side_offset = 0,
  only_inside = FALSE
)

create_raster_hexagonal(
  geometry,
  cell_size,
  side_offset = 0,
  only_inside = FALSE
)
}
\arguments{
\item{geometry}{\code{\link[sf]{sf}} \code{data.frame} containing geometry which should be cover by
the grid.}

\item{cell_size}{\code{numeric} specifing the distance for equal spaced points or cells.}

\item{side_offset}{\code{numeric} specifing the side offset, distance added to bounding box
to generate points for KDE. Good estimate is usually the same value as band width of KDE.}

\item{only_inside}{\code{logical} specifing if the points should be generated only inside of the
geometry. Only valid for \code{"POLYGON"} or \code{"MULTIPOLYGON"} types. Default value is
\code{FALSE}.}
}
\value{
\code{\link[sf]{sf}} \code{data.frame}.
}
\description{
Create raster of equaly spaced rectangles or hexagons. The distance between centre points
in both x and y dimension is equal to \code{cell_size}.
}
\section{Functions}{
\itemize{
\item \code{create_raster_rectangular}: Create rectangular raster

\item \code{create_raster_hexagonal}: Create hexagonal raster
}}

